LoB v0.10.0
Added:
Shield powerup! Pick it up to be immune from damage for 3 seconds.
You can now reload the game with the R button!
Added a nested array that contains all enemies. Added enemy counter on top of the screen.

Changed:
Cleaned the code up in general.
Removed old unused files.
Changed som of the size of text, and added some more text to the How To Play screen.

Fixed:
Fixed a bug with where unpausing and pausing again while a powerup is active made the powerup run out while paused.
Re-added the favicon.