LoB v0.11.0
Added:
Functions to spawn the different kinds of powerups.
Added a maximum amount of max health upgrades.
Enemy counter that displays the amount of enemies left on the map.
Multiplier that affects the % of powerups spawning based on the enemy type.

Changed:
Changed the way powerups and healthpacks spawn. Instead of being on a timer, they now have a certain % chance of spawning when you kill an enemy. This prevents stalling for healthpacks while there is only one enemy alive.
Changed the way the game handles the player being dead.
Changed the max amount of damage upgrades from 2 to 1.